requirements.txt
zeep
mysql-connector-python
python-dotenv